pyproject.toml
[project]
name = "langchain-rag-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain==1.2.*",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
]
